backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4

slowapi==0.1.9
sqlalchemy==1.4.49
python-jose==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5

numpy==1.26.4
pandas==2.2.1
pyyaml==6.0.1
requests==2.31.0
python-multipart==0.0.9
openpyxl==3.1.2

gspread==6.0.2
oauth2client==4.1.3
google-cloud-bigquery==3.17.2

groq==0.37.1
redis==5.2.1
pyodbc==5.1.0
psycopg2-binary==2.9.9
snowflake-connector-python==3.7.2